7475 Extract Onboarding MainContentView
