+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/Onboarding Errors/InPersonPaymentsOnboardingError.swift
@@ -4,7 +4,7 @@ import Yosemite
 struct InPersonPaymentsOnboardingError: View {
     let title: String
     let message: String
-    let image: ImageInfo
+    let image: InPersonPaymentsOnboardingErrorMainContentView.ImageInfo
     let supportLink: Bool
     let learnMore: Bool
     var button: ButtonInfo? = nil
@@ -14,16 +14,11 @@ struct InPersonPaymentsOnboardingError: View {
         let action: () -> Void
     }
 
-    struct ImageInfo {
-        let image: UIImage
-        let height: CGFloat
-    }
-
     var body: some View {
         VStack {
             Spacer()
 
-            MainContent(
+            InPersonPaymentsOnboardingErrorMainContentView(
                 title: title,
                 message: message,
                 image: image,
@@ -41,41 +36,6 @@ struct InPersonPaymentsOnboardingError: View {
                 InPersonPaymentsLearnMore()
             }
         }.padding()
-    }
-
-    struct MainContent: View {
-        let title: String
-        let message: String
-        let image: ImageInfo
-        let supportLink: Bool
-
-        @Environment(\.verticalSizeClass) var verticalSizeClass
-
-        var isCompact: Bool {
-            get {
-                verticalSizeClass == .compact
-            }
-        }
-
-        var body: some View {
-            VStack(alignment: .center) {
-                Text(title)
-                    .font(.headline)
-                    .padding(.bottom, isCompact ? 16 : 32)
-                Image(uiImage: image.image)
-                    .resizable()
-                    .scaledToFit()
-                    .frame(height: isCompact ? image.height / 3 : image.height)
-                    .padding(.bottom, isCompact ? 16 : 32)
-                Text(message)
-                    .font(.callout)
-                    .padding(.bottom, isCompact ? 12 : 24)
-                if supportLink {
-                    InPersonPaymentsSupportLink()
-                }
-            }.multilineTextAlignment(.center)
-            .frame(maxWidth: 500)
-        }
     }
 }
