+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/Onboarding Errors/InPersonPaymentsPluginConflictAdminView.swift
@@ -15,10 +15,10 @@ struct InPersonPaymentsPluginConflictAdmin: View {
         ScrollableVStack {
             Spacer()
 
-            InPersonPaymentsOnboardingError.MainContent(
+            InPersonPaymentsOnboardingErrorMainContentView(
                 title: Localization.title,
                 message: Localization.message,
-                image: InPersonPaymentsOnboardingError.ImageInfo(
+                image: InPersonPaymentsOnboardingErrorMainContentView.ImageInfo(
                     image: .paymentErrorImage,
                     height: 108.0
                 ),
