+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/Onboarding Errors/InPersonPaymentsPluginNotSetup.swift
@@ -10,10 +10,10 @@ struct InPersonPaymentsPluginNotSetup: View {
         ScrollableVStack {
             Spacer()
 
-            InPersonPaymentsOnboardingError.MainContent(
+            InPersonPaymentsOnboardingErrorMainContentView(
                 title: String(format: Localization.title, plugin.pluginName),
                 message: String(format: Localization.message, plugin.pluginName),
-                image: InPersonPaymentsOnboardingError.ImageInfo(
+                image: InPersonPaymentsOnboardingErrorMainContentView.ImageInfo(
                     image: plugin.image,
                     height: 108.0
                 ),
